Moved configuration files to config package
