+++ src/main/java/com/example/jwtapp/JwtAppApplication.java
@@ -2,8 +2,6 @@ package com.example.jwtapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @SpringBootApplication
 public class JwtAppApplication {
@@ -11,11 +9,4 @@ public class JwtAppApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(JwtAppApplication.class, args);
 	}
-	
-
-	@Bean
-	public BCryptPasswordEncoder encoder() {
-		return new BCryptPasswordEncoder();
-	}
-
 }
