+++ src/main/java/com/example/jwtapp/config/AuthorizationServerConfig.java
@@ -1,4 +1,4 @@
-package com.example.jwtapp;
+package com.example.jwtapp.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
